breaking: Renames libs to src
