+++ src/HotelManager.js
