+++ test/BookingData.js
@@ -10,9 +10,9 @@ const moment = require('moment');
 const Web3 = require('web3');
 const provider = new Web3.providers.HttpProvider('http://localhost:8545');
 const web3 = new Web3(provider);
-const web3providerFactory = require('../libs/web3provider');
+const web3providerFactory = require('../src/web3provider');
 
-const BookingData = require('../libs/BookingData');
+const BookingData = require('../src/BookingData');
 const help = require('./helpers/index');
 
 describe('BookingData', function () {
