+++ test/HotelEvents.js
@@ -1,4 +1,4 @@
-const User = require('../libs/User');
+const User = require('../src/User');
 const help = require('./helpers/index');
 
 const assert = require('chai').assert;
@@ -6,9 +6,9 @@ const assert = require('chai').assert;
 const Web3 = require('web3');
 const provider = new Web3.providers.HttpProvider('http://localhost:8545');
 const web3 = new Web3(provider);
-const web3providerFactory = require('../libs/web3provider');
+const web3providerFactory = require('../src/web3provider');
 
-const HotelEvents = require('../libs/HotelEvents.js');
+const HotelEvents = require('../src/HotelEvents.js');
 
 xdescribe('HotelEvents', function () {
   let Manager;
