+++ test/User.js
@@ -7,9 +7,9 @@ const sinon = require('sinon');
 const Web3 = require('web3');
 const provider = new Web3.providers.HttpProvider('http://localhost:8545');
 const web3 = new Web3(provider);
-const web3providerFactory = require('../libs/web3provider');
+const web3providerFactory = require('../src/web3provider');
 
-const User = require('../libs/User');
+const User = require('../src/User');
 const help = require('./helpers/index');
 
 describe('User', function () {
